Ensure error code is an int
While Exception->getCode is typed to return an int, it's still not
guaranteed to.

See: https://www.php.net/manual/en/exception.getcode.php

> Returns the exception code as int in Exception but possibly as other
> type in Exception descendants (for example as string in PDOException).

So that we do not inadvertently throw a TypeError, ensure the code is an
int before passing it into the constructor. String codes will still be
accessible via the previous exception.